git-svn-id: file:///p/wpis/people/students/kidd/svnwali/trunk@295 dfa36f5b-b2fa-0310-9605-e5aca7e64462

+++ Source/wali/witness/VisitorPrinter.hpp
@@ -33,10 +33,10 @@ namespace wali {
 
             protected:
                 std::ostream& os;
-                int depth;
+                size_t depth;
 
             private:
-                VisitorPrinter( std::ostream& os, int depth );
+                VisitorPrinter( std::ostream& os, size_t depth );
         };
 
     } // namespace witness
